fix: correções após merge
