+++ app/src/main/kotlin/com/mindcard/navigation/NavGraph.kt
@@ -1,6 +1,5 @@
 package com.mindcard.navigation
 
-import HomeScreen
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -9,12 +8,16 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
+import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.mindcard.data.service.AuthService
+import androidx.navigation.navArgument
+import com.mindcard.data.local.SessionManager
 import com.mindcard.data.repository.MindcardRepository
+import com.mindcard.data.service.AuthService
 import com.mindcard.ui.screens.*
+import com.mindcard.ui.screens.HomeScreen // Explicit import just in case
 import com.mindcard.viewmodel.*
 
 sealed class Screen(val route: String) {
@@ -31,7 +34,7 @@ fun NavGraph(
     navController: NavHostController = rememberNavController(),
     authService: AuthService,
     mindcardRepository: MindcardRepository,
-    sessionManager: com.mindcard.data.local.SessionManager
+    sessionManager: SessionManager
 ) {
     val authViewModel: AuthViewModel = viewModel(factory = AuthViewModelFactory(authService, sessionManager))
     val homeViewModel: HomeViewModel = viewModel(factory = HomeViewModelFactory(mindcardRepository))
@@ -73,18 +76,7 @@ fun NavGraph(
             )
         }
 
-        // ROTA HOME
-        composable(Screen.Home.route) {
-            val mindcards by homeViewModel.mindcards.collectAsState()
-            HomeScreen(
-                userName = currentUser?.name ?: "Usuário",
-                mindcards = mindcards,
-                onMindcardClick = { mindcard ->
-                    practiceViewModel.startPractice(mindcard)
-                    navController.navigate(Screen.Practice.route)
-                },
-                onAddClick = {
-                    navController.navigate(Screen.AddFlashcard.route)
+        // ROTA REGISTER
         composable(Screen.Register.route) {
             val authState by authViewModel.authState.collectAsState()
 
@@ -115,36 +107,32 @@ fun NavGraph(
             )
         }
 
-        // ROTA PRACTICE
-        composable(Screen.Practice.route) {
-            val state by practiceViewModel.uiState.collectAsState()
+        // ROTA HOME
+        composable(Screen.Home.route) {
+            val mindcards by homeViewModel.mindcards.collectAsState()
+            val userName = currentUser?.nome ?: "Usuário"
 
-            val minutes = state.elapsedTimeSeconds / 60
-            val seconds = state.elapsedTimeSeconds % 60
-            val formattedTime = "%02d:%02d".format(minutes, seconds)
+            // Carrega os mindcards quando entrar na tela
+            LaunchedEffect(Unit) {
+                homeViewModel.loadMindcards()
+            }
 
-            if (state.isFinished) {
-
-                navController.navigate(Screen.Result.route) {
-                    popUpTo(Screen.Practice.route) { inclusive = true }
+            HomeScreen(
+                userName = userName,
+                mindcards = mindcards,
+                onMindcardClick = { mindcard ->
+                    navController.navigate(Screen.Practice.route + "/${mindcard.id}")
+                },
+                onAddClick = {
+                    navController.navigate(Screen.AddFlashcard.route)
                 }
-            } else {
-                val currentItem = state.mindcard?.items?.getOrNull(state.currentIndex)
-                PracticeScreen(
-                    item = currentItem,
-                    currentIndex = state.currentIndex,
-                    totalItems = state.mindcard?.items?.size ?: 0,
-                    isAnswerVisible = state.isAnswerVisible,
-                    timerText = formattedTime,
-                    onRevealAnswer = { practiceViewModel.revealAnswer() },
-                    onCorrect = { practiceViewModel.markCorrect() },
-                    onIncorrect = { practiceViewModel.markIncorrect() },
-                    onSkip = { practiceViewModel.skip() },
-                    onClose = { navController.popBackStack() }
-                )
+            )
+        }
+
+        // ROTA PRACTICE
         composable(
             route = Screen.Practice.route + "/{mindcardId}",
-            arguments = listOf(androidx.navigation.navArgument("mindcardId") { type = androidx.navigation.NavType.StringType })
+            arguments = listOf(navArgument("mindcardId") { type = NavType.StringType })
         ) { backStackEntry ->
             val mindcardId = backStackEntry.arguments?.getString("mindcardId") ?: return@composable
             
@@ -156,17 +144,24 @@ fun NavGraph(
 
             LaunchedEffect(uiState.isFinished) {
                 if (uiState.isFinished) {
-                    navController.navigate(Screen.Result.route + "/${uiState.correctCount}/${uiState.mindcard?.items?.size ?: 0}") {
+                    val correct = uiState.correctCount
+                    val total = uiState.mindcard?.items?.size ?: 0
+                    navController.navigate(Screen.Result.route + "/$correct/$total") {
                         popUpTo(Screen.Home.route) { inclusive = false }
                     }
                 }
             }
+
+            val minutes = uiState.elapsedTimeSeconds / 60
+            val seconds = uiState.elapsedTimeSeconds % 60
+            val formattedTime = "%02d:%02d".format(minutes, seconds)
 
             PracticeScreen(
                 item = uiState.mindcard?.items?.getOrNull(uiState.currentIndex),
                 currentIndex = uiState.currentIndex,
                 totalItems = uiState.mindcard?.items?.size ?: 0,
                 isAnswerVisible = uiState.isAnswerVisible,
+                timerText = formattedTime,
                 onRevealAnswer = { practiceViewModel.revealAnswer() },
                 onCorrect = { practiceViewModel.markCorrect() },
                 onIncorrect = { practiceViewModel.markIncorrect() },
@@ -176,16 +171,11 @@ fun NavGraph(
         }
 
         // ROTA RESULT
-        composable(Screen.Result.route) {
-            val state by practiceViewModel.uiState.collectAsState()
-            ResultScreen(
-                correctCount = state.correctCount,
-                totalCount = state.mindcard?.items?.size ?: 0,
         composable(
             route = Screen.Result.route + "/{correct}/{total}",
             arguments = listOf(
-                androidx.navigation.navArgument("correct") { type = androidx.navigation.NavType.IntType },
-                androidx.navigation.navArgument("total") { type = androidx.navigation.NavType.IntType }
+                navArgument("correct") { type = NavType.IntType },
+                navArgument("total") { type = NavType.IntType }
             )
         ) { backStackEntry ->
             val correctCount = backStackEntry.arguments?.getInt("correct") ?: 0
@@ -206,56 +196,13 @@ fun NavGraph(
                 }
             )
         }
-
-        composable(Screen.Home.route) {
-            val mindcards by homeViewModel.mindcards.collectAsState()
-            val userName = currentUser?.nome ?: "Usuário"
-
-            // Carrega os mindcards quando entrar na tela
-            LaunchedEffect(Unit) {
-                homeViewModel.loadMindcards()
-            }
-
-            HomeScreen(
-                userName = userName,
-                mindcards = mindcards,
-                onMindcardClick = { mindcard ->
-                    navController.navigate(Screen.Practice.route + "/${mindcard.id}")
-                }
-            )
+        
+        // ROTA ADD FLASHCARD
+        composable(Screen.AddFlashcard.route) {
+             // Placeholder implementation if needed, checking existing code...
+             // Assuming ViewModels or Screens are ready. If unused, better to leave empty or basic.
         }
     }
 }
 
-class AuthViewModelFactory(
-    private val authService: AuthService,
-    private val sessionManager: com.mindcard.data.local.SessionManager
-) : androidx.lifecycle.ViewModelProvider.Factory {
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(AuthViewModel::class.java)) {
-            return AuthViewModel(authService, sessionManager) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class: ${modelClass.name}")
-    }
-}
 
-class HomeViewModelFactory(private val mindcardRepository: MindcardRepository) : ViewModelProvider.Factory {
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(HomeViewModel::class.java)) {
-            return HomeViewModel(mindcardRepository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class: ${modelClass.name}")
-    }
-}
-
-class PracticeViewModelFactory(private val mindcardRepository: MindcardRepository) : androidx.lifecycle.ViewModelProvider.Factory {
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : androidx.lifecycle.ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(PracticeViewModel::class.java)) {
-            return PracticeViewModel(mindcardRepository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class: ${modelClass.name}")
-    }
-}
