+++ app/src/main/kotlin/com/mindcard/ui/screens/HomeScreen.kt
@@ -1,3 +1,5 @@
+package com.mindcard.ui.screens
+
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
